# Speed up scores_to_quintiles 800 times