Results from my machine:
```python

In [2]: %%time
   ...: scores_to_quintiles_2(x)
   ...: 
   ...: 
CPU times: user 43.3 ms, sys: 12.5 ms, total: 55.8 ms
Wall time: 54 ms
Out[2]: 
array([[0, 3, 4, ..., 1, 2, 4],
       [2, 0, 2, ..., 0, 1, 0],
       [4, 3, 4, ..., 0, 4, 3],
       ...,
       [4, 4, 4, ..., 0, 0, 1],
       [2, 3, 4, ..., 1, 2, 3],
       [3, 4, 4, ..., 2, 4, 2]])

In [3]: %%time
   ...: scores_to_quintiles(x)
   ...: 
   ...: 
CPU times: user 42.9 s, sys: 3.25 ms, total: 42.9 s
Wall time: 42.9 s
Out[3]: 
array([[0, 3, 4, ..., 1, 2, 4],
       [2, 0, 2, ..., 0, 1, 0],
       [4, 3, 4, ..., 0, 4, 3],
       ...,
       [4, 4, 4, ..., 0, 0, 1],
       [2, 3, 4, ..., 1, 2, 3],
       [3, 4, 4, ..., 2, 4, 2]])
```

(in the above, `scores_to_quintiles_2` is the version I'm submitting in this PR)

`q` doesn't depend on what's being iterated over, so computing it outside the loop drastically speeds things up